removed old question data, created axios post to use decisionCode to get decision, updated with navigation with new menu items
