+++ frontend/src/components/landing/Navigation.css
@@ -6,6 +6,7 @@
   justify-content: space-between;
   width: 1000px;
   padding-top: 25px;
+  padding-bottom: 20px;
 }
 
 /* logo */
