+++ frontend/src/components/question/Decision.css
@@ -5,6 +5,11 @@
 
 .decision-title {
   font-size: 30px;
+  padding-top: 20px;
+  padding-bottom: 10px;
+}
+
+.decision-code {
   padding-bottom: 30px;
 }
